added lockRemaining
