+++ contracts/CurveConvexStrategyV2Native.sol
@@ -10,12 +10,12 @@ import "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
 import "@openzeppelin/contracts-upgradeable/utils/AddressUpgradeable.sol";
 import "@openzeppelin/contracts-upgradeable/utils/structs/EnumerableSetUpgradeable.sol";
 
-import {IAlluoStrategyV2} from "../../../interfaces/IAlluoStrategyV2.sol";
+import {IAlluoStrategyV2} from "./interfaces/IAlluoStrategyV2.sol";
 import {ICvxBooster} from "./interfaces/ICvxBooster.sol";
 import {ICvxBaseRewardPool} from "./interfaces/ICvxBaseRewardPool.sol";
-import {IExchange} from "../../../interfaces/IExchange.sol";
-import {IPriceFeedRouterV2} from "../../../interfaces/IPriceFeedRouterV2.sol";
-import {IWrappedEther} from "../../../interfaces/IWrappedEther.sol";
+import {IExchange} from "./interfaces/IExchange.sol";
+import {IPriceFeedRouterV2} from "./interfaces/IPriceFeedRouterV2.sol";
+import {IWrappedEther} from "./interfaces/IWrappedEther.sol";
 
 contract CurveConvexStrategyV2Native is
     IAlluoStrategyV2,
@@ -520,7 +520,7 @@ contract CurveConvexStrategyV2Native is
         upgradeStatus = _status;
     }
 
-    function _authorizeUpgrade(address)
+    function _authorizeUpgrade(address newImplementation)
         internal
         override
         onlyRole(UPGRADER_ROLE)
