+++ test/CurveConvexStrategyV2Native.ts
