+++ test/IntegratedFraxConvex.ts
@@ -288,10 +288,10 @@ describe("Automated strategy execution", function () {
             const executorBalanceBefore = await poolToken.balanceOf(executor.address);
             const rq1 = await executor.callStatic.encodeLiquidityCommand(_codeName, 6000);
             const rq2 = await executor.callStatic.encodeLiquidityCommand("Curve/Convex Mim+3CRV", 4000);
-            const rq3 = await executor.callStatic.encodeLiquidityCommand("Curve/Convex Mim+3CRV", 0);
-            const rq4 = await executor.callStatic.encodeLiquidityCommand(_codeName, 0);
+            // const rq3 = await executor.callStatic.encodeLiquidityCommand("Curve/Convex Mim+3CRV", 0);
+            // const rq4 = await executor.callStatic.encodeLiquidityCommand(_codeName, 0);
 
-            const encodedMmessages = await executor.callStatic.encodeAllMessages([rq1[0], rq2[0], rq3[0], rq4[0]], [rq1[1], rq2[1], rq3[1], rq4[1]]);
+            const encodedMmessages = await executor.callStatic.encodeAllMessages([rq1[0], rq2[0]], [rq1[1], rq2[1]]);
             const inputData = encodedMmessages[2];
             await executor.submitData(inputData);
